fix(emu): correctly calculate carry flag for subtraction (and cmp)

+++ mmvm.go
@@ -1749,19 +1749,19 @@ func (cpu *CPU) Step(inst Instruction) {
 			b := cpu.Get8(s2)
 			r := a - b
 			cpu.Set8(s1, r)
-			cpu.Flags().SetZSCO(r == 0, r < 0, a < b, r > math.MaxInt8 || r < math.MinInt8)
+			cpu.Flags().SetZSCO(r == 0, r < 0, uint16(a) < uint16(b), r > math.MaxInt8 || r < math.MinInt8)
 		case s1.W() == 1 && s2.W() == 1:
 			a := cpu.Get8(s1)
 			b := cpu.Get8(s2)
 			r := a - b
 			cpu.Set16(s1, r)
-			cpu.Flags().SetZSCO(r == 0, r < 0, a < b, r > math.MaxInt16 || r < math.MinInt16)
+			cpu.Flags().SetZSCO(r == 0, r < 0, uint16(a) < uint16(b), r > math.MaxInt16 || r < math.MinInt16)
 		case s1.W() == 1 && s2.W() == 0 && isImm(s2):
 			a := cpu.Get8(s1)
 			b := cpu.Get8(s2)
 			r := a - b
 			cpu.Set8(s1, r)
-			cpu.Flags().SetZSCO(r == 0, r < 0, a < b, r > math.MaxInt16 || r < math.MinInt16)
+			cpu.Flags().SetZSCO(r == 0, r < 0, uint16(a) < uint16(b), r > math.MaxInt16 || r < math.MinInt16)
 		}
 	case OpSsbRegRm, OpSsbRmImm, OpSsbAccImm:
 	case OpDecRm, OpDecReg:
@@ -1776,17 +1776,17 @@ func (cpu *CPU) Step(inst Instruction) {
 			a := cpu.Get8(s1)
 			b := cpu.Get8(s2)
 			r := a - b
-			cpu.Flags().SetZSCO(r == 0, r < 0, a < b, r > math.MaxInt8 || r < math.MinInt8)
+			cpu.Flags().SetZSCO(r == 0, r < 0, uint16(a) < uint16(b), r > math.MaxInt8 || r < math.MinInt8)
 		case s1.W() == 1 && s2.W() == 1:
 			a := cpu.Get8(s1)
 			b := cpu.Get8(s2)
 			r := a - b
-			cpu.Flags().SetZSCO(r == 0, r < 0, a < b, r > math.MaxInt16 || r < math.MinInt16)
+			cpu.Flags().SetZSCO(r == 0, r < 0, uint16(a) < uint16(b), r > math.MaxInt16 || r < math.MinInt16)
 		case s1.W() == 1 && s2.W() == 0 && isImm(s2):
 			a := cpu.Get8(s1)
 			b := cpu.Get8(s2)
 			r := a - b
-			cpu.Flags().SetZSCO(r == 0, r < 0, a < b, r > math.MaxInt16 || r < math.MinInt16)
+			cpu.Flags().SetZSCO(r == 0, r < 0, uint16(a) < uint16(b), r > math.MaxInt16 || r < math.MinInt16)
 		}
 	case OpAAS:
 	case OpDAS:
